requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3

# Database - Using SQLite locally (uncomment for PostgreSQL)
# psycopg2-binary==2.9.9
# dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
cryptography==41.0.5
python-decouple==3.8

# AI/ML Libraries
openai==1.3.5
anthropic==0.7.2
google-generativeai==0.3.1

# Payment Integrations
stripe==7.5.0
paypalrestsdk==1.13.1
razorpay==1.3.0

# Email & SMS
sendgrid==6.11.0
twilio==8.10.1


# Background Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Utilities
Pillow==10.1.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Development
python-dotenv==1.0.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
